Refactor LogPublisher to infrastructure package and update logging implementation
